refactor: MagicBook에서 Page와 type을 분리
