+++ src/pages/MagicBook/types.ts
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export type PageProps = {
+  front: React.ReactNode;
+  back: React.ReactNode;
+  page: number;
+};
